#6080: rename create_group_norm_input_mask
