+++ models/experimental/functional_stable_diffusion/tt2/ttnn_functional_resnetblock2d.py
@@ -257,7 +257,7 @@ class resnetBlock2D:
                 ttnn.to_torch(self.parameters.norm1.bias), in_channels, num_cores_across_channel
             )
 
-            self.norm1_input_mask_torch_tensor = ttnn.create_groupnorm_input_mask(
+            self.norm1_input_mask_torch_tensor = ttnn.create_group_norm_input_mask(
                 in_channels, self.groups, num_cores_across_channel
             )
 
@@ -317,7 +317,7 @@ class resnetBlock2D:
                 memory_config=ttnn.DRAM_MEMORY_CONFIG,
             )
 
-            self.norm2_input_mask_torch_tensor = ttnn.create_groupnorm_input_mask(
+            self.norm2_input_mask_torch_tensor = ttnn.create_group_norm_input_mask(
                 out_channels, self.groups, num_cores_across_channel
             )
             self.norm2_input_mask = ttnn.from_torch(
