+++ models/experimental/functional_stable_diffusion/tt2/ttnn_functional_transformer_2d.py
@@ -112,7 +112,7 @@ class transformer_2d_model:
                 memory_config=ttnn.DRAM_MEMORY_CONFIG,
             )
 
-            self.norm_input_mask_torch_tensor = ttnn.create_groupnorm_input_mask(
+            self.norm_input_mask_torch_tensor = ttnn.create_group_norm_input_mask(
                 in_channels, norm_num_groups, num_cores_across_channel
             )
             self.norm_input_mask = ttnn.from_torch(
