+++ ttnn/ttnn/__init__.py
@@ -327,8 +327,7 @@ from ttnn.operations.normalization import (
     rms_norm,
     group_norm,
     create_group_norm_weight_bias_rm,
-    create_groupnorm_input_mask,
-    manual_group_norm,
+    create_group_norm_input_mask,
     determine_expected_group_norm_sharded_config_and_grid_size,
 )
